app/server_requirements.txt
torch
opencv-python
uvicorn
fastapi
Pillow
torchvision
python-multipart